+++ Sources/ZenVoice/SettingsWindowController.swift
@@ -21,6 +21,7 @@ final class SettingsWindowController: NSObject, NSWindowDelegate {
     private let viewModel: SettingsViewModel
     private let historyViewModel: HistoryViewModel
     private let audioHistoryViewModel: AudioHistoryViewModel
+    private let meetingViewModel: MeetingViewModel
     private let cloudAIViewModel: CloudAIViewModel
     private let updatesViewModel: UpdatesViewModel
     private let insightsViewModel: InsightsViewModel
@@ -33,6 +34,7 @@ final class SettingsWindowController: NSObject, NSWindowDelegate {
         viewModel: SettingsViewModel,
         historyViewModel: HistoryViewModel,
         audioHistoryViewModel: AudioHistoryViewModel,
+        meetingViewModel: MeetingViewModel,
         cloudAIViewModel: CloudAIViewModel,
         updatesViewModel: UpdatesViewModel,
         insightsViewModel: InsightsViewModel,
@@ -45,6 +47,7 @@ final class SettingsWindowController: NSObject, NSWindowDelegate {
         self.viewModel = viewModel
         self.historyViewModel = historyViewModel
         self.audioHistoryViewModel = audioHistoryViewModel
+        self.meetingViewModel = meetingViewModel
         self.cloudAIViewModel = cloudAIViewModel
         self.updatesViewModel = updatesViewModel
         self.insightsViewModel = insightsViewModel
@@ -89,6 +92,7 @@ final class SettingsWindowController: NSObject, NSWindowDelegate {
                 viewModel: viewModel,
                 historyViewModel: historyViewModel,
                 audioHistoryViewModel: audioHistoryViewModel,
+                meetingViewModel: meetingViewModel,
                 cloudAIViewModel: cloudAIViewModel,
                 updatesViewModel: updatesViewModel,
                 insightsViewModel: insightsViewModel,